Reorganize space utils
