+++ static/src/containers/Spaces/reducer.js
@@ -2,43 +2,9 @@ import { REMOVE_SPACE } from './actions';
 import { UPDATE_MAP_POSITION } from '../Map/actions';
 import { FETCH_SOUNDS_REQUEST, FETCH_SOUNDS_SUCCESS, FETCH_SOUNDS_FAILURE }
   from '../Sounds/actions';
-import { computeSoundGlobalPosition } from '../Sounds/utils';
-import { getMapCenter } from '../Map/utils';
-import { generateListOfSpacesOriginPositions } from './utils';
+import { computeSpacePosition, computeSpacePositionInMap, computeSpaceIndex,
+  getClosestSpaceToCenter } from './utils';
 import sessions from '../Sessions/reducer';
-
-const computeSpacePosition = (spaceIndex) => {
-  const { x, y } = generateListOfSpacesOriginPositions(spaceIndex)[spaceIndex];
-  return { x: (4 * x) + 1, y: (4 * y) + 1 };
-};
-
-const computeSpacePositionInMap = (spacePosition, mapPosition) => {
-  const tsnePosition = { x: 0, y: 0 };
-  const { cx, cy } = computeSoundGlobalPosition(tsnePosition, spacePosition, mapPosition);
-  return {
-    x: cx,
-    y: cy,
-  };
-};
-
-const computeSpaceIndex = (spaces) => {
-  // We find the smallest non-negative integer which is not already
-  // taken as an index of any of existing spaces
-  const usedSpaceIndexes = [];
-  spaces.forEach((space) => {
-    usedSpaceIndexes.push(space.spaceIndex);
-  });
-  let spaceIndex = 0;
-  let found = false;
-  while (!found) {
-    if (!usedSpaceIndexes.includes(spaceIndex)) {
-      found = true;
-    } else {
-      spaceIndex += 1;
-    }
-  }
-  return spaceIndex;
-};
 
 const spaceInitialState = {
   sounds: [],
@@ -113,23 +79,6 @@ const spacesReducer = (state = [], action) => {
     default:
       return state;
   }
-};
-
-const getSpaceDistanceToCenter = (space, center) =>
-  Math.sqrt(Math.pow((space.currentPositionInMap.x - center.x), 2) +
-    Math.pow((space.currentPositionInMap.y - center.y), 2));
-
-const getClosestSpaceToCenter = (allSpaces) => {
-  let minDistance = Infinity;
-  const mapCenter = getMapCenter();
-  return allSpaces.reduce((curState, curSpace) => {
-    const distanceToCenter = getSpaceDistanceToCenter(curSpace, mapCenter);
-    if (distanceToCenter < minDistance) {
-      minDistance = distanceToCenter;
-      return curSpace;
-    }
-    return curState;
-  }, undefined);
 };
 
 const currentSpace = (state = '', action, allSpaces) => {
